moving home page to feature
