+++ src/features/home/animated-description/animated-description.component.tsx
