+++ src/features/home/animated-description/index.ts
