+++ src/features/icon/components/icon.component.css
