+++ src/features/icon/components/icon.component.tsx
