+++ src/features/icon/store/icon.types.ts
