+++ src/features/icon/store/index.ts
