+++ src/features/shell/components/footer/footer.component.css
