+++ src/features/shell/components/footer/footer.component.tsx
