+++ src/features/shell/components/footer/index.ts
