+++ src/features/shell/store/shell.reducer.ts
