+++ src/features/shell/store/shell.types.ts
