+++ src/pages/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import { Home } from './home';
-import { ShellContainer } from '../features/+shell';
+import { Home } from '../features/home';
+import { ShellContainer } from '../features/shell';
 
 class SiteIndex extends React.Component {
   render() {
